// apps/web/src/components/editor/sidebar/scriptSideBar/types/ActionTypes.ts
export enum ActionType {
  // Transform Actions (Node2D)
  MOVE_TO = 'move_to',
  MOVE_BY = 'move_by',
  TRANSLATE = 'translate',
  ROTATE_TO = 'rotate_to',
  ROTATE_BY = 'rotate_by',
  ROTATE = 'rotate',
  SCALE_TO = 'scale_to',
  SCALE_BY = 'scale_by',
  LOOK_AT = 'look_at',
  GET_ANGLE_TO = 'get_angle_to',
  TO_LOCAL = 'to_local',
  TO_GLOBAL = 'to_global',
  
  // Node Core Functions
  ADD_CHILD = 'add_child',
  REMOVE_CHILD = 'remove_child',
  QUEUE_FREE = 'queue_free',
  GET_NODE = 'get_node',
  GET_PARENT = 'get_parent',
  GET_CHILDREN = 'get_children',
  SET_PROCESS = 'set_process',
  SET_PHYSICS_PROCESS = 'set_physics_process',
  
  // Sprite2D Actions
  SET_TEXTURE = 'set_texture',
  GET_TEXTURE = 'get_texture',
  SET_REGION_ENABLED = 'set_region_enabled',
  SET_REGION_RECT = 'set_region_rect',
  SET_FLIP_H = 'set_flip_h',
  SET_FLIP_V = 'set_flip_v',
  
  // AnimatedSprite2D Actions
  PLAY_ANIMATION = 'play_animation',
  STOP_ANIMATION = 'stop_animation',
  PAUSE_ANIMATION = 'pause_animation',
  SET_ANIMATION = 'set_animation',
  GET_ANIMATION = 'get_animation',
  SET_FRAME = 'set_frame',
  GET_FRAME = 'get_frame',
  SET_SPEED_SCALE = 'set_speed_scale',
  
  // RigidBody2D Actions
  APPLY_IMPULSE = 'apply_impulse',
  APPLY_FORCE = 'apply_force',
  SET_GRAVITY_SCALE = 'set_gravity_scale',
  FREEZE_BODY = 'freeze_body',
  SET_FREEZE_ENABLED = 'set_freeze_enabled',
  SET_MASS = 'set_mass',
  SET_LINEAR_VELOCITY = 'set_linear_velocity',
  SET_ANGULAR_VELOCITY = 'set_angular_velocity',
  SET_LOCK_ROTATION = 'set_lock_rotation',
  
  // StaticBody2D Actions
  MOVE_AND_COLLIDE = 'move_and_collide',
  SET_CONSTANT_LINEAR_VELOCITY = 'set_constant_linear_velocity',
  SET_CONSTANT_ANGULAR_VELOCITY = 'set_constant_angular_velocity',
  
  // Area2D Actions
  GET_OVERLAPPING_BODIES = 'get_overlapping_bodies',
  GET_OVERLAPPING_AREAS = 'get_overlapping_areas',
  OVERLAPS_BODY = 'overlaps_body',
  OVERLAPS_AREA = 'overlaps_area',
  SET_MONITORING = 'set_monitoring',
  SET_MONITORABLE = 'set_monitorable',
  SET_GRAVITY = 'set_gravity',
  SET_GRAVITY_VECTOR = 'set_gravity_vector',
  
  // CharacterBody2D Actions
  MOVE_AND_SLIDE = 'move_and_slide',
  IS_ON_FLOOR = 'is_on_floor',
  IS_ON_CEILING = 'is_on_ceiling',
  IS_ON_WALL = 'is_on_wall',
  GET_FLOOR_NORMAL = 'get_floor_normal',
  SET_VELOCITY = 'set_velocity',
  SET_FLOOR_MAX_ANGLE = 'set_floor_max_angle',
  SET_UP_DIRECTION = 'set_up_direction',
  
  // CollisionShape2D Actions
  SET_SHAPE = 'set_shape',
  GET_SHAPE = 'get_shape',
  SET_COLLISION_DISABLED = 'set_collision_disabled',
  SET_ONE_WAY_COLLISION = 'set_one_way_collision',
  
  // CollisionPolygon2D Actions
  SET_POLYGON = 'set_polygon',
  GET_POLYGON = 'get_polygon',
  SET_BUILD_MODE = 'set_build_mode',
  
  // Property Actions
  SET_VISIBLE = 'set_visible',
  SET_MODULATE = 'set_modulate',
  SET_PROPERTY = 'set_property',
  
  // Animation Actions
  TWEEN_PROPERTY = 'tween_property',
  
  // Signal Actions
  EMIT_SIGNAL = 'emit_signal',
  CONNECT_SIGNAL = 'connect_signal',
  
  // Flow Control
  WAIT = 'wait',
  CONDITION = 'condition',
  LOOP = 'loop',
  
  // Scene Actions
  CHANGE_SCENE = 'change_scene',
  INSTANTIATE_SCENE = 'instantiate_scene',
  
  // Audio Actions
  PLAY_SOUND = 'play_sound',
  STOP_SOUND = 'stop_sound',
  
  // Input Actions
  CHECK_INPUT = 'check_input',
  
  // Debug Actions
  PRINT = 'print',
  
  // Advanced
  CUSTOM_CODE = 'custom_code'
}

// Specific Action Interfaces
export interface MoveToAction extends BaseAction {
  type: ActionType.MOVE_TO;
  targetX: number;
  targetY: number;
  duration?: number;
  easeType?: EaseType;
}

export interface MoveByAction extends BaseAction {
  type: ActionType.MOVE_BY;
  deltaX: number;
  deltaY: number;
  duration?: number;
  easeType?: EaseType;
}

export interface RotateToAction extends BaseAction {
  type: ActionType.ROTATE_TO;
  targetRotation: number;
  duration?: number;
  easeType?: EaseType;
}

export interface RotateByAction extends BaseAction {
  type: ActionType.ROTATE_BY;
  deltaRotation: number;
  duration?: number;
  easeType?: EaseType;
}

export interface ScaleToAction extends BaseAction {
  type: ActionType.SCALE_TO;
  targetScaleX: number;
  targetScaleY: number;
  duration?: number;
  easeType?: EaseType;
}

export interface SetVisibleAction extends BaseAction {
  type: ActionType.SET_VISIBLE;
  visible: boolean;
}

export interface SetModulateAction extends BaseAction {
  type: ActionType.SET_MODULATE;
  color: string;
  alpha: number;
}

export interface SetPropertyAction extends BaseAction {
  type: ActionType.SET_PROPERTY;
  propertyPath: string;
  value: string;
  valueType: PropertyValueType;
}

export interface TweenPropertyAction extends BaseAction {
  type: ActionType.TWEEN_PROPERTY;
  propertyPath: string;
  targetValue: string;
  duration: number;
  easeType: EaseType;
  transitionType: TransitionType;
}

export interface EmitSignalAction extends BaseAction {
  type: ActionType.EMIT_SIGNAL;
  signalName: string;
  parameters: SignalParameter[];
}

export interface WaitAction extends BaseAction {
  type: ActionType.WAIT;
  duration: number;
  waitType: WaitType;
}

export interface ConditionAction extends BaseAction {
  type: ActionType.CONDITION;
  leftOperand: string;
  operator: ComparisonOperator;
  rightOperand: string;
  trueActions: string[]; // Action IDs
  falseActions: string[]; // Action IDs
}

export interface PrintAction extends BaseAction {
  type: ActionType.PRINT;
  message: string;
  includeNodeName: boolean;
}

export interface PlaySoundAction extends BaseAction {
  type: ActionType.PLAY_SOUND;
  soundPath: string;
  volume: number;
  pitch: number;
  loop: boolean;
}

export interface CustomCodeAction extends BaseAction {
  type: ActionType.CUSTOM_CODE;
  code: string;
}

// New Godot Action Interfaces

// Node2D Transform Actions
export interface TranslateAction extends BaseAction {
  type: ActionType.TRANSLATE;
  vector: { x: number; y: number };
}

export interface RotateAction extends BaseAction {
  type: ActionType.ROTATE;
  angle: number;
}

export interface LookAtAction extends BaseAction {
  type: ActionType.LOOK_AT;
  point: { x: number; y: number };
}

export interface GetAngleToAction extends BaseAction {
  type: ActionType.GET_ANGLE_TO;
  point: { x: number; y: number };
}

export interface ToLocalAction extends BaseAction {
  type: ActionType.TO_LOCAL;
  globalPoint: { x: number; y: number };
}

export interface ToGlobalAction extends BaseAction {
  type: ActionType.TO_GLOBAL;
  localPoint: { x: number; y: number };
}

// Node Core Actions
export interface AddChildAction extends BaseAction {
  type: ActionType.ADD_CHILD;
  childPath: string;
}

export interface RemoveChildAction extends BaseAction {
  type: ActionType.REMOVE_CHILD;
  childPath: string;
}

export interface GetNodeAction extends BaseAction {
  type: ActionType.GET_NODE;
  nodePath: string;
}

export interface SetProcessAction extends BaseAction {
  type: ActionType.SET_PROCESS;
  enabled: boolean;
}

export interface SetPhysicsProcessAction extends BaseAction {
  type: ActionType.SET_PHYSICS_PROCESS;
  enabled: boolean;
}

// Sprite2D Actions
export interface SetTextureAction extends BaseAction {
  type: ActionType.SET_TEXTURE;
  texturePath: string;
}

export interface SetRegionEnabledAction extends BaseAction {
  type: ActionType.SET_REGION_ENABLED;
  enabled: boolean;
}

export interface SetRegionRectAction extends BaseAction {
  type: ActionType.SET_REGION_RECT;
  x: number;
  y: number;
  width: number;
  height: number;
}

export interface SetFlipHAction extends BaseAction {
  type: ActionType.SET_FLIP_H;
  flip: boolean;
}

export interface SetFlipVAction extends BaseAction {
  type: ActionType.SET_FLIP_V;
  flip: boolean;
}

// AnimatedSprite2D Actions
export interface StopAnimationAction extends BaseAction {
  type: ActionType.STOP_ANIMATION;
}

export interface PauseAnimationAction extends BaseAction {
  type: ActionType.PAUSE_ANIMATION;
}

export interface SetAnimationAction extends BaseAction {
  type: ActionType.SET_ANIMATION;
  animationName: string;
}

export interface SetFrameAction extends BaseAction {
  type: ActionType.SET_FRAME;
  frame: number;
}

export interface SetSpeedScaleAction extends BaseAction {
  type: ActionType.SET_SPEED_SCALE;
  speedScale: number;
}

// RigidBody2D Actions
export interface ApplyImpulseAction extends BaseAction {
  type: ActionType.APPLY_IMPULSE;
  impulse: { x: number; y: number };
  position?: { x: number; y: number };
}

export interface ApplyForceAction extends BaseAction {
  type: ActionType.APPLY_FORCE;
  force: { x: number; y: number };
  position?: { x: number; y: number };
}

export interface SetGravityScaleAction extends BaseAction {
  type: ActionType.SET_GRAVITY_SCALE;
  scale: number;
}

export interface FreezeBodyAction extends BaseAction {
  type: ActionType.FREEZE_BODY;
}

export interface SetFreezeEnabledAction extends BaseAction {
  type: ActionType.SET_FREEZE_ENABLED;
  enabled: boolean;
}

export interface SetMassAction extends BaseAction {
  type: ActionType.SET_MASS;
  mass: number;
}

export interface SetLinearVelocityAction extends BaseAction {
  type: ActionType.SET_LINEAR_VELOCITY;
  velocity: { x: number; y: number };
}

export interface SetAngularVelocityAction extends BaseAction {
  type: ActionType.SET_ANGULAR_VELOCITY;
  velocity: number;
}

export interface SetLockRotationAction extends BaseAction {
  type: ActionType.SET_LOCK_ROTATION;
  locked: boolean;
}

// StaticBody2D Actions
export interface MoveAndCollideAction extends BaseAction {
  type: ActionType.MOVE_AND_COLLIDE;
  velocity: { x: number; y: number };
}

export interface SetConstantLinearVelocityAction extends BaseAction {
  type: ActionType.SET_CONSTANT_LINEAR_VELOCITY;
  velocity: { x: number; y: number };
}

export interface SetConstantAngularVelocityAction extends BaseAction {
  type: ActionType.SET_CONSTANT_ANGULAR_VELOCITY;
  velocity: number;
}

// Area2D Actions
export interface GetOverlappingBodiesAction extends BaseAction {
  type: ActionType.GET_OVERLAPPING_BODIES;
}

export interface GetOverlappingAreasAction extends BaseAction {
  type: ActionType.GET_OVERLAPPING_AREAS;
}

export interface OverlapsBodyAction extends BaseAction {
  type: ActionType.OVERLAPS_BODY;
  bodyPath: string;
}

export interface OverlapsAreaAction extends BaseAction {
  type: ActionType.OVERLAPS_AREA;
  areaPath: string;
}

export interface SetMonitoringAction extends BaseAction {
  type: ActionType.SET_MONITORING;
  enabled: boolean;
}

export interface SetMonitorableAction extends BaseAction {
  type: ActionType.SET_MONITORABLE;
  enabled: boolean;
}
// ...
export type Action = 
  | MoveToAction 
  | MoveByAction 
  | RotateToAction 
  | RotateByAction 
  | ScaleToAction
  | SetVisibleAction
  | SetModulateAction
  | SetPropertyAction
  | TweenPropertyAction
  | EmitSignalAction
  | WaitAction
  | ConditionAction
  | PrintAction
  | PlaySoundAction
  | CustomCodeAction
  // New Godot Actions
  | TranslateAction
  | RotateAction
  | LookAtAction
  | GetAngleToAction
  | ToLocalAction
  | ToGlobalAction
  | AddChildAction
  | RemoveChildAction
  | GetNodeAction
  | SetProcessAction
  | SetPhysicsProcessAction
  | SetTextureAction
  | SetRegionEnabledAction
  | SetRegionRectAction
  | SetFlipHAction
  | SetFlipVAction
  | StopAnimationAction
  | PauseAnimationAction
  | SetAnimationAction
  | SetFrameAction
  | SetSpeedScaleAction
  | ApplyImpulseAction
  | ApplyForceAction
  | SetGravityScaleAction
  | FreezeBodyAction
  | SetFreezeEnabledAction
  | SetMassAction
  | SetLinearVelocityAction
  | SetAngularVelocityAction
  | SetLockRotationAction
  | MoveAndCollideAction
  | SetConstantLinearVelocityAction
  | SetConstantAngularVelocityAction
  | GetOverlappingBodiesAction
  | GetOverlappingAreasAction
  | OverlapsBodyAction
  | OverlapsAreaAction
  | SetMonitoringAction
  | SetMonitorableAction
  | SetGravityAction
  | SetGravityVectorAction
  | MoveAndSlideAction
  | IsOnFloorAction
  | IsOnCeilingAction
  | IsOnWallAction
  | GetFloorNormalAction
  | SetVelocityAction
  | SetFloorMaxAngleAction
  | SetUpDirectionAction
  | SetShapeAction
  | SetCollisionDisabledAction
  | SetOneWayCollisionAction
  | SetPolygonAction
  | SetBuildModeAction;